refactor: clear stuff related to AuthController
