--- user-service/src/main/java/org/linkwave/userservice/dto/AuthDto.java
@@ -1,15 +1,0 @@
-package org.linkwave.userservice.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-@EqualsAndHashCode
-public class AuthDto {
-    private String jwtAccess;
-    private String jwtRefresh;
-}
