--- user-service/src/main/java/org/linkwave/userservice/dto/UserRefreshRequest.java
@@ -1,18 +1,0 @@
-package org.linkwave.userservice.dto;
-
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-@NoArgsConstructor
-@AllArgsConstructor
-@Getter
-public class UserRefreshRequest {
-
-    @NotNull(message = "can't be null")
-    @NotBlank(message = "can't be blank")
-    private String jwtRefresh;
-
-}
